#Authorization screens and logic #Send-receive screens and logic #Add coin screen

+++ entities/src/main/java/com/wowlet/entities/local/ActivityItem.kt
@@ -1,3 +1,3 @@
 package com.wowlet.entities.local
 
-data class ActivityItem(var icon:String, val name:String, val symbolsPrice:String, val price:Double, val lamports:Double, val slot:Double, val signature:String, val fee:Double, val from:String, val to:String,val tokenName:String)
+data class ActivityItem(var icon:String, val name:String, val symbolsPrice:String, val price:Double, val lamports:Double, val slot:Long, val signature:String, val fee:Double, val from:String, val to:String,val tokenName:String)
